+++ package.json
@@ -26,6 +26,7 @@
     "lodash": "4.17.20",
     "moment": "2.29.1",
     "react-i18next": "11.11.0",
+    "react-query": "3.17.1",
     "react-relay": "11.0.1",
     "react-simple-maps": "2.3.0",
     "relay-runtime": "11.0.1",
@@ -62,8 +63,7 @@
     "react-dom": "17.0.1",
     "react-scripts": "4.0.1",
     "prop-types": "15.7.2",
-    "axios": "0.21.1",
-    "react-query": "3.17.1",
+    "axios": "0.21.1",    
     "@material-ui/lab": "4.0.0-alpha.57"
   },
   "devDependencies": {
