+++ src/hooks/businessObjects/index.js
@@ -1,0 +1,7 @@
+export {
+    default as useUserContacts,
+} from './useUserContacts'
+
+export {
+    default as useUserLocales,
+} from './useUserLocales'
