+++ src/hooks/businessObjects/useUserContacts.js
@@ -1,0 +1,8 @@
+import getContacts from './getContacts'
+import useQueryGetContacts from './useQueryGetContacts'
+
+const useUserContacts = () => {
+    return useQueryGetContacts(getContacts)
+}
+
+export default useUserContacts
